Conclusion : Need to make a third person controller
To create a Bow character movement.
->
I watched a Brackeys tutorial and imported the Sci fi Weapons Asset.
->
After Importing I got to know that I need to import an asset but in the newer version of unity it was not there.
->
So I went to Unity Asset Store and Found it there
->
I saw First Person and Third Person, Third Person is what I need so I have kept in asset manager but not imported
->
Now I don't need that much functionality, so I will Try to make my own third person controller and than make the bow work , umm somewhat.